Make requested changes, move V2 components to analytics and fix comment defs.
