+++ assets/js/modules/analytics/components/dashboard/Version2/DashboardAllTrafficWidgetV2.js
@@ -1,5 +1,5 @@
 /**
- * DashboardAllTrafficWidget component, mainly used for screen-reader text.
+ * DashboardAllTrafficWidget component
  *
  * Site Kit by Google, Copyright 2020 Google LLC
  *
@@ -26,14 +26,15 @@ import { _x } from '@wordpress/i18n';
  */
 import Data from 'googlesitekit-data';
 import Widgets from 'googlesitekit-widgets';
-import { STORE_NAME as MODULE_ANALYTICS, FORM_ALL_TRAFFIC_WIDGET } from '../../../modules/analytics/datastore/constants';
-import { STORE_NAME as CORE_FORMS } from '../../../googlesitekit/datastore/forms/constants';
-import whenActive from '../../../util/when-active';
+import { STORE_NAME as MODULE_ANALYTICS, FORM_ALL_TRAFFIC_WIDGET } from '../../../datastore/constants';
+import { STORE_NAME as CORE_FORMS } from '../../../../../googlesitekit/datastore/forms/constants';
+import whenActive from '../../../../../util/when-active';
 import TotalUserCount from './TotalUserCount';
 import UserCountGraph from './UserCountGraph';
 import DimensionTabs from './DimensionTabs';
 import UserDimensionsPieChart from './UserDimensionsPieChart';
-import SourceLink from '../../SourceLink';
+import SourceLink from '../../../../../components/SourceLink';
+import { Grid, Row, Cell } from '../../../../../material-components/layout';
 const { Widget } = Widgets.components;
 const { useSelect } = Data;
 
@@ -42,31 +43,33 @@ function DashboardAllTrafficWidget() {
 	const dimensionValue = useSelect( ( select ) => select( CORE_FORMS ).getValue( FORM_ALL_TRAFFIC_WIDGET, 'dimensionValue' ) );
 	const serviceReportURL = useSelect( ( select ) => select( MODULE_ANALYTICS ).getServiceReportURL( 'trafficsources-overview' ) );
 
-	return <Widget
-		slug="analyticsAllTrafficV2"
-		noPadding
-		footer={ () => (
-			<SourceLink
-				className="googlesitekit-data-block__source"
-				name={ _x( 'Analytics', 'Service name', 'google-site-kit' ) }
-				href={ serviceReportURL }
-				external
-			/>
-		) }
-	>
-		<div className="mdc-layout-grid">
-			<div className="mdc-layout-grid__inner">
-				<div className="mdc-layout-grid__cell mdc-layout-grid__cell--span-8">
-					<TotalUserCount dimensionName={ dimensionName } dimensionValue={ dimensionValue } />
-					<UserCountGraph dimensionName={ dimensionName } dimensionValue={ dimensionValue } />
-				</div>
-				<div className="mdc-layout-grid__cell mdc-layout-grid__cell--span-4">
-					<DimensionTabs />
-					<UserDimensionsPieChart />
-				</div>
-			</div>
-		</div>
-	</Widget>;
+	return (
+		<Widget
+			slug="analyticsAllTrafficV2"
+			noPadding
+			footer={ () => (
+				<SourceLink
+					className="googlesitekit-data-block__source"
+					name={ _x( 'Analytics', 'Service name', 'google-site-kit' ) }
+					href={ serviceReportURL }
+					external
+				/>
+			) }
+		>
+			<Grid>
+				<Row>
+					<Cell size={ 8 }>
+						<TotalUserCount dimensionName={ dimensionName } dimensionValue={ dimensionValue } />
+						<UserCountGraph dimensionName={ dimensionName } dimensionValue={ dimensionValue } />
+					</Cell>
+					<Cell size={ 4 }>
+						<DimensionTabs />
+						<UserDimensionsPieChart />
+					</Cell>
+				</Row>
+			</Grid>
+		</Widget>
+	);
 }
 
 export default whenActive( { moduleName: 'analytics' } )( DashboardAllTrafficWidget );
