+++ assets/js/modules/analytics/components/dashboard/Version2/DimensionTabs.js
@@ -1,5 +1,5 @@
 /**
- * UserCountGraph component, mainly used for screen-reader text.
+ * DimensionTabs component
  *
  * Site Kit by Google, Copyright 2020 Google LLC
  *
@@ -16,19 +16,6 @@
  * limitations under the License.
  */
 
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-
-// eslint-disable-next-line no-unused-vars
-function UserCountGraph( { dimensionName, dimensionValue } ) {
+export default function DimensionTabs() {
 	return null;
 }
-
-UserCountGraph.propTypes = {
-	dimensionName: PropTypes.string,
-	dimensionValue: PropTypes.string,
-};
-
-export default UserCountGraph;
